@@ -11,29 +11,30 @@
     "test:ci": "meteor test --once --driver-package dispatch:mocha-phantomjs"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "5.0.5",
+    "@angular/compiler-cli": "6.1.10",
     "@types/chai": "4.0.8",
     "@types/meteor": "1.4.12",
     "@types/mocha": "2.2.44",
     "@types/sinon": "4.0.0",
     "chai": "4.1.2",
+    "chai-spies": "1.0.0",
+    "meteor-node-stubs": "0.4.1",
     "phantomjs-prebuilt": "2.1.16",
     "sinon": "4.1.2",
-    "typescript": "2.6.2",
-    "chai-spies": "1.0.0",
-    "meteor-node-stubs": "0.4.1"
+    "typescript": "2.9.2"
   },
   "dependencies": {
-    "@angular/animations": "5.0.5",
+    "@angular/animations": "6.1.10",
     "@angular/cdk": "5.0.0-rc0",
-    "@angular/common": "5.0.5",
-    "@angular/compiler": "5.0.5",
-    "@angular/core": "5.0.5",
-    "@angular/forms": "5.0.5",
-    "@angular/http": "5.0.5",
-    "@angular/platform-browser": "5.0.5",
-    "@angular/platform-browser-dynamic": "5.0.5",
-    "@angular/router": "5.0.5",
+    "@angular/cli": "^6.2.9",
+    "@angular/common": "6.1.10",
+    "@angular/compiler": "6.1.10",
+    "@angular/core": "6.1.10",
+    "@angular/forms": "6.1.10",
+    "@angular/http": "6.1.10",
+    "@angular/platform-browser": "6.1.10",
+    "@angular/platform-browser-dynamic": "6.1.10",
+    "@angular/router": "6.1.10",
     "@babel/runtime": "^7.13.10",
     "angular2-csv": "0.2.5",
     "angular2-infinite-scroll": "^0.3.5",
@@ -51,8 +52,8 @@
     "ng2-charts": "^1.6.0",
     "papaparse": "^5.3.0",
     "reflect-metadata": "0.1.12",
-    "rxjs": "5.5.2",
-    "rxjs-compat": "^6.2.1",
-    "zone.js": "0.8.18"
+    "rxjs": "6.6.7",
+    "rxjs-compat": "^6.6.7",
+    "zone.js": "0.8.29"
   }
 }
