@@ -1,12 +1,11 @@
 import {
 	Observable
-} from 'rxjs/Observable';
+, 
+	Subject
+} from 'rxjs';
 import {
 	Injectable
 } from '@angular/core';
-import {
-	Subject
-} from 'rxjs/Subject';
 @Injectable()
 export class SharedNavigationService {
 	// Observable string sources
