@@ -7,10 +7,9 @@ import {
 } from 'meteor/meteor';
 import {
     Observable
-} from 'rxjs/Observable';
-import {
+, 
     Subscription
-} from 'rxjs/Subscription';
+} from 'rxjs';
 import {
     MongoObservable
 } from 'meteor-rxjs';
