@@ -20,10 +20,9 @@ import {
 } from 'meteor/meteor';
 import {
 	Observable
-} from 'rxjs/Observable';
-import {
+, 
 	Subscription
-} from 'rxjs/Subscription';
+} from 'rxjs';
 import {
 	MeteorObservable
 } from 'meteor-rxjs';
