@@ -17,10 +17,9 @@ import {
 } from 'meteor/meteor';
 import {
 	Observable
-} from 'rxjs/Observable';
-import {
+, 
 	Subscription
-} from 'rxjs/Subscription';
+} from 'rxjs';
 import {
 	Roles
 } from 'meteor/alanning:roles';
